Create README.md
Background: 
Our client, a sports analyst, approached us with an exciting prospect: to predict the outcome of NBA games using just four statistics from each team. The concept was revolutionary and had the potential to change the way the game was viewed. However, there were some concerns as the team compositions changed frequently, and the four data points for each team would only be available at the end of each game, which meant that they would already know the result by that time.

Problem: 
The issue was that we needed to find a way to predict the four data points for each team from the previous games they had played to develop a time series analysis. This would then be fed into our machine-learning model. We explained our reservations to the client regarding using data from the previous season as a predictor for the current season, which would ultimately impact the accuracy of the predictions.

Solution: 
After much discussion, we decided to use state-of-the-art ARIMA modelling and to enable the code to choose the best hyperparameters to get the minimum mean squared error (MSE) score. The four data points generated by the ARIMA model were then fed into a separate algorithm for data pre-processing and used to predict the probability of each team winning using Logistic Regression, XGBoost, and Random Forest models.
Our approach yielded an amazing level of accuracy of 90% or higher each time we ran the model with different sets of teams. We advised the client that the accuracy would decrease when predicting results for the current season with limited data points. Therefore, we suggested using a voting system and going with the outcome suggested by at least two models. The client was pleased with the results and offered to conduct training programs for other sports analysts on how to use data science in the world of sports.